Rework how we cleanup child nodes
We need to start cleaning the last node first so we don't remove child dirs while we
are iterating over the children.

Given this:
Test
  - test1
  - test2
Test2
  - test21
  - test22

If we cleanup Test then numberOfChildren() will be decremented to 1 and
Test2 will never get cleaned up but plugin will be unloaded so Test2 branch nodes
will be garbage.